// src/lib/database.types.ts
export interface Database {
  public: {
    Tables: {
// ...
      accounts: {
        Row: {
          balance: number;
          created_at: string;
          icon_name: string;
          id: string;
          name: string;
          type: string;
          user_id: string;
          workspace_id: string;
        };
        Insert: {
          balance: number;
          created_at?: string;
          icon_name: string;
          id?: string;
          name: string;
          type: string;
          user_id: string;
          workspace_id: string;
        };
        Update: {
          balance?: number;
          created_at?: string;
          icon_name?: string;
          id?: string;
          name?: string;
          type?: string;
          user_id?: string;
          workspace_id?: string;
        };
        Relationships: [
          {
            foreignKeyName: "accounts_user_id_fkey";
            columns: ["user_id"];
            isOneToOne: false;
            referencedRelation: "users";
            referencedColumns: ["id"];
          },
          {
            foreignKeyName: "accounts_workspace_id_fkey";
            columns: ["workspace_id"];
            isOneToOne: false;
            referencedRelation: "workspaces";
            referencedColumns: ["id"];
          }
        ];
      };
// ...
      users: {
        Row: {
          auto_transfer_savings: boolean;
          currency: string;
          default_expense_account_id: string | null;
          full_name: string | null;
          id: string;
          secondary_currency: string | null;
          updated_at: string;
        };
        Insert: {
          auto_transfer_savings?: boolean;
          currency: string;
          default_expense_account_id?: string | null;
          full_name?: string | null;
          id: string;
          secondary_currency?: string | null;
          updated_at?: string;
        };
        Update: {
          auto_transfer_savings?: boolean;
          currency?: string;
          default_expense_account_id?: string | null;
          full_name?: string | null;
          id?: string;
          secondary_currency?: string | null;
          updated_at?: string;
        };
        Relationships: [
          {
            foreignKeyName: "users_default_expense_account_id_fkey";
            columns: ["default_expense_account_id"];
            isOneToOne: false;
            referencedRelation: "accounts";
            referencedColumns: ["id"];
          }
        ];
      };
// ...
    };
// ...
  };
}